+++ config/toml.go
@@ -289,14 +289,10 @@ size = {{ .Mempool.Size }}
 # Limit the total size of all txs in the mempool.
 # This only accounts for raw transactions (e.g. given 1MB transactions and
 # max_txs_bytes=5MB, mempool will only accept 5 transactions).
-max_txs_bytes = {{ .Mempool.MaxTxsBytes }}
+max_pending_txs_bytes = {{ .Mempool.MaxPendingTxsBytes }}
 
 # Size of the cache (used to filter transactions we saw earlier) in transactions
 cache_size = {{ .Mempool.CacheSize }}
-
-# Maximum size of a single transaction.
-# NOTE: the max size of a tx transmitted over the network is {max_tx_bytes} + {amino overhead}.
-max_tx_bytes = {{ .Mempool.MaxTxBytes }}
 
 ##### fast sync configuration options #####
 [fastsync]