+++ mempool/errors.go
@@ -13,8 +13,8 @@ var (
 
 // ErrTxTooLarge means the tx is too big to be sent in a message to other peers
 type ErrTxTooLarge struct {
-	max    int
-	actual int
+	max    int64
+	actual int64
 }
 
 func (e ErrTxTooLarge) Error() string {
@@ -35,19 +35,4 @@ func (e ErrMempoolIsFull) Error() string {
 		"mempool is full: number of txs %d (max: %d), total txs bytes %d (max: %d)",
 		e.numTxs, e.maxTxs,
 		e.txsBytes, e.maxTxsBytes)
-}
-
-// ErrPreCheck is returned when tx is too big
-type ErrPreCheck struct {
-	Reason error
-}
-
-func (e ErrPreCheck) Error() string {
-	return e.Reason.Error()
-}
-
-// IsPreCheckError returns true if err is due to pre check failure.
-func IsPreCheckError(err error) bool {
-	_, ok := err.(ErrPreCheck)
-	return ok
 }